Style the error page and add more information about the error and what can be done.
