+++ src/views.rs
@@ -28,41 +28,4 @@ pub async fn corpora(State(state): State<Arc<GlobalAppState>>) -> Result<impl In
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::tests::get_html;
-    use axum::{body::Body, http::Request};
-    use mockito::mock;
-    use scraper::Selector;
-    use tower::ServiceExt;
-    use tracing_test::traced_test;
-
-    #[tokio::test]
-    #[traced_test]
-    async fn list_corpora() {
-        let m = mock("GET", "/corpora")
-            .with_header("content-type", "application/json")
-            .with_body(r#"["pcc2", "demo.dialog"]"#)
-            .create();
-        {
-            let app = crate::app().unwrap();
-
-            let response = app
-                .oneshot(Request::builder().uri("/").body(Body::empty()).unwrap())
-                .await
-                .unwrap();
-
-            assert_eq!(response.status(), StatusCode::OK);
-
-            let html = get_html(response).await;
-            let list_selector = Selector::parse(".content > ul > li").unwrap();
-            let corpora: Vec<_> = html
-                .select(&list_selector)
-                .map(|e| e.text().collect::<Vec<_>>().join(""))
-                .collect();
-
-            assert_eq!(vec!["demo.dialog", "pcc2"], corpora);
-        }
-        m.assert();
-    }
-}
+mod tests;
