+++ templates/base.html
@@ -37,7 +37,7 @@
         </div>
       </div>
     </nav>
-    <div class="content">
+    <div class="content box">
       {% block content %}
       {% endblock content %}
     </div>
